refactor(#14) Moves DatabaseAttribute and ItemMapper to their own module
