+++ tests/test_database/__init__.py
@@ -1,2 +1,3 @@
 from .test_engine import TestDatabaseEngine
 from .test_item import TestDatabaseItem
+from .test_attribute import TestDatabaseAttribute
